requirements.txt
alembic==1.7.3
Babel==2.9.1
blinker==1.4
Bootstrap-Flask==1.8.0
cachelib==0.4.1
cffi==1.14.6
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.2
Flask-Babel==2.0.0
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-MySQLdb==0.2.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
Mako==1.1.5
MarkupSafe==2.0.1
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.0
numpy==1.21.4
pandas==1.3.4
patsy==0.5.2
plotly==5.4.0
plspm==0.5.6
pycparser==2.20
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
SQLAlchemy==1.4.27
statsmodels==0.13.1
tenacity==8.0.1
threadpoolctl==3.0.0
visitor==0.1.3
Werkzeug==2.0.1
WTForms==3.0.0
XlsxWriter==3.0.2
